Refactor and move Extension and ExtensionType
